Create a separate file for concat test
